+++ src/IO/ADIOS/ADIOS2Auxiliary.cpp
@@ -93,6 +93,9 @@ namespace detail
             { "float", Datatype::FLOAT },
             { "double", Datatype::DOUBLE },
             { "long double", Datatype::LONG_DOUBLE },
+            { "float complex", Datatype::CFLOAT },
+            { "double complex", Datatype::CDOUBLE },
+            { "long double complex", Datatype::CLONG_DOUBLE }, // does not exist as of 2.6.0 but might come later
             { "uint8_t", Datatype::UCHAR },
             { "int8_t", Datatype::CHAR },
             { "uint16_t", determineDatatype< uint16_t >() },